classification-proxy-server/requirements.txt
Flask==1.0.2
pandas==0.23.4
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
oauth2client==4.1.3